requirements.txt
google-adk
fastapi
uvicorn
python-dotenv
 